refactor: extract useGroupingSort, useCascadeFilters, useSetupLanding, useSavedViewsManager from WorksCalendar
Phase 2 of WorksCalendar.tsx refactor. Moves ~370 lines of inline state/effects/callbacks
into four focused custom hooks. WorksCalendar.tsx drops from ~2598 to ~2230 lines with
no behaviour changes — all 2643 tests pass.

- src/core/setupRecipes.ts: pure util moved from WorksCalendar module scope
- src/hooks/useGroupingSort.ts: activeGroupBy/Sort/ShowAllGroups + sidebarGroupLevels
- src/hooks/useCascadeFilters.ts: cascade selections ↔ filter pipeline sync
- src/hooks/useSetupLanding.ts: setup dismissed/skip/reopen/finish
- src/hooks/useSavedViewsManager.ts: saved view apply/dirty-track/delete/sidebar-save
